requirements.txt
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
regex>=2024.5.15
scikit-learn>=1.3.0
rank-bm25>=0.2.2
pyarrow>=15.0.0
# Optional (if you want embeddings); code will fallback if unavailable
sentence-transformers>=3.0.0
torch>=2.2.0
# Optional OCR (only if you set parse.ocr_if_needed: true)
Pillow>=10.0.0
pytesseract>=0.3.10

requests>=2.31.0

pinecone-client>=4.1.0
